cxr-pneumonia/requirements.txt
# Core
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.24
pandas>=2.0
pillow>=10.0
scikit-learn>=1.3
tqdm>=4.66
PyYAML>=6.0

# Visualization & Explainability
matplotlib>=3.7
seaborn>=0.12
opencv-python>=4.8
pytorch-grad-cam>=1.5.0

# App
streamlit>=1.30
